+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * **Breaking change:** Calling any function that produces an error in the headless client will now raise an exception instead of returning the old format of `{"success": bool, "message": str}`. Applications should catch these exceptions to handle these errors and display the error message to the user.
   * Commands that run successfully will simply return their success message as a `str`.
   * Commands that don't print any messages on success will return `None`.
+* The `restart` function has been temporarily disabled due to a bug in the headless client preventing its use. Please see [this GitHub issue](https://github.com/Neos-Metaverse/NeosPublic/issues/1841) for more information. Calling `restart` will raise a `NotImplementedError` exception for now.
 
 ## 2021-05-28
 * Added `host` and `port` attributes to the `RemoteHeadlessClient` class.